#76: Added validation to ensure data submissions are only allowed to the category assigned to the study (#94)
* feat: validate study submissions against assigned category

* Changes related to the feedback

* chore: adding Copyrights to new files

* fix: remove studyId param from submission API and update lookup logic

* refactor(submission): simplify study/category validation and clean up error messages

* Update src/controllers/submission/submit.ts


* Update src/controllers/submission/submit.ts


* Update src/controllers/submission/submit.ts


* Validate study organization during category check

* fix(submission): update validation to throw error if study name mismatches organization

* fix(submission): enforce study_name and categoryId validation in editData

* fix(submission): correct study_name comparison in validation check

* fix(submission): standardize validation and error message in submit/editData

* fix(submission): standardize study not found validation and message

---------
